requirements.txt
django
mysqlclient
mysql-connector-python
Pillow
easy-thumbnails
django-ckeditor